third_party/blink/web_tests/fast/events/touch/gesture/tap-target-matches-active-expected.txt

Verifies that the element receiving the :active style is the same as the element receiving the click event, even in the presence of difficult touch adjustment scenarios

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Sending gestureTapDown
PASS getHoverActiveState(target) is "default"
Sending gestureShowPress
PASS getHoverActiveState(target) is "hoveredAndActive"
Sending gestureTap
PASS clickTarget.id is "target"
PASS successfullyParsed is true

TEST COMPLETE
